backend/requirements.txt
# Web framework
flask

# Google Cloud SDKs
google-cloud-bigquery
google-cloud-core

python-dotenv

